sample_data/pyproject.toml
[project]
name = "sample-data"
version = "0.1.0"
description = "Exports data from the full Elasticsearch index to a smaller sample index that can be used for development and testing purposes."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "elasticsearch>=8.0.0,<9",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
line-length = 100
output-format = "grouped"
show-fixes = true

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
